Update html
I was able to update the code provided by simon. I swapped out all of the wording on each of the pages to suit my own website, I then replaced the entire Simon game with a chat box. It isn't functioning properly yet, but i was able to put in the HTML that allowed the user to type in a box and then the text gets inserted in the HTML where it would display under the previous message. It doesnt save, and several aspects are hard coded in place as a placeholder jsut so I can visualize the final product better. I also removed the score page as I will not be needing it.