Create about.css
I realised that my HTML was a big mess. I deciced to start over with the new downloaded code that implemets the CSS. I just finished the about page and got it to a point where I am comfortable. I needed toadd some padding around the outside to make it look nicer. I also needed to add new rules. I replaced the image that was there with an AI generated one. Ill replace it again later