Create index.css
I added a few things to the Simon css for the index page. I didn't like how the box to the left of the email and password were not alighted so i made sure to align them. I ended up making the text in those boxes not aligned so I still need to work that out