Create play.css
Here is all the CSS I had created. I previously put some CSS on the HTML page but I moved it onto this document. I refitted 3 collumns to the page: my pals, the chat, and the options collumns. I managed to get it to look nice on a computer screen but still need to mess with it to be better fitted for a phone screen.